refactor: migrate vod reviews and dashboard
- Migrate VOD Reviews (2 controllers)
- Migrate Dashboard module
- Organize video review features
